Feature/unit and integration test

+++ src/main/java/org/juliantovar/arquitecturahexagonal/infrastructure/dto/response/weatherapi/CurrentDto.java
@@ -21,7 +21,7 @@ public record CurrentDto(
         @JsonProperty("wind_kph")
         Double windSpeed,
         @JsonProperty("pressure_mb")
-        Double pressure,
+        Integer pressure,
         int humidity,
         @JsonProperty("feelslike_c")
         Double feelsLike,
